Refactor test app deploy annotations
