+++ systemtests/src/main/java/io/enmasse/systemtest/annotations/ExternalClientsExtension.java
@@ -2,7 +2,7 @@
  * Copyright 2019, EnMasse authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.enmasse.systemtest.messagingclients;
+package io.enmasse.systemtest.annotations;
 
 
 import io.enmasse.systemtest.platform.apps.SystemtestsKubernetesApps;
