+++ systemtests/src/main/java/io/enmasse/systemtest/annotations/SeleniumFirefox.java
@@ -2,7 +2,7 @@
  * Copyright 2019, EnMasse authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.enmasse.systemtest.messagingclients;
+package io.enmasse.systemtest.annotations;
 
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -14,6 +14,6 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 @Target({ElementType.METHOD, ElementType.TYPE, ElementType.ANNOTATION_TYPE})
 @Retention(RUNTIME)
-@ExtendWith(ExternalClientsExtension.class)
-public @interface ExternalClients {
+@ExtendWith(SeleniumFirefoxExtension.class)
+public @interface SeleniumFirefox {
 }
