+++ systemtests/src/main/java/io/enmasse/systemtest/annotations/SeleniumFirefoxExtension.java
@@ -2,9 +2,11 @@
  * Copyright 2019, EnMasse authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.enmasse.systemtest.selenium;
+package io.enmasse.systemtest.annotations;
 
 import io.enmasse.systemtest.Environment;
+import io.enmasse.systemtest.selenium.SeleniumManagement;
+import io.enmasse.systemtest.selenium.SeleniumProvider;
 import io.enmasse.systemtest.utils.TestUtils;
 import org.junit.jupiter.api.extension.AfterAllCallback;
 import org.junit.jupiter.api.extension.AfterTestExecutionCallback;
@@ -12,19 +14,19 @@ import org.junit.jupiter.api.extension.BeforeAllCallback;
 import org.junit.jupiter.api.extension.BeforeTestExecutionCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 
-public class SeleniumChromeExtension implements BeforeTestExecutionCallback, AfterTestExecutionCallback, BeforeAllCallback, AfterAllCallback {
+public class SeleniumFirefoxExtension implements BeforeTestExecutionCallback, AfterTestExecutionCallback, BeforeAllCallback, AfterAllCallback {
     private boolean isFullClass = false;
 
     @Override
     public void afterAll(ExtensionContext extensionContext) throws Exception {
         SeleniumProvider.getInstance().tearDownDrivers();
-        SeleniumManagement.removeChromeApp();
+        SeleniumManagement.removeFirefoxApp();
         isFullClass = false;
     }
 
     @Override
     public void beforeAll(ExtensionContext extensionContext) throws Exception {
-        SeleniumManagement.deployChromeApp();
+        SeleniumManagement.deployFirefoxApp();
         isFullClass = true;
     }
 
@@ -35,7 +37,7 @@ public class SeleniumChromeExtension implements BeforeTestExecutionCallback, Aft
         }
         SeleniumProvider.getInstance().tearDownDrivers();
         if (!isFullClass) {
-            SeleniumManagement.removeChromeApp();
+            SeleniumManagement.removeFirefoxApp();
         } else {
             SeleniumManagement.restartSeleniumApp();
         }
@@ -44,10 +46,10 @@ public class SeleniumChromeExtension implements BeforeTestExecutionCallback, Aft
     @Override
     public void beforeTestExecution(ExtensionContext extensionContext) throws Exception {
         if (!isFullClass) {
-            SeleniumManagement.deployChromeApp();
+            SeleniumManagement.deployFirefoxApp();
         }
         if (SeleniumProvider.getInstance().getDriver() == null)
-            SeleniumProvider.getInstance().setupDriver(TestUtils.getChromeDriver());
+            SeleniumProvider.getInstance().setupDriver(TestUtils.getFirefoxDriver());
         else
             SeleniumProvider.getInstance().clearScreenShots();
     }
