+++ systemtests/src/test/java/io/enmasse/systemtest/framework/SupportedInstallTypesConditionTest.java
@@ -2,7 +2,7 @@
  * Copyright 2019, EnMasse authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.enmasse.systemtest.condition;
+package io.enmasse.systemtest.framework;
 
 import io.enmasse.systemtest.EnmasseInstallType;
 import io.enmasse.systemtest.OLMInstallationType;
@@ -10,6 +10,8 @@ import io.enmasse.systemtest.TestTag;
 
 import java.lang.annotation.Annotation;
 
+import io.enmasse.systemtest.condition.SupportedInstallType;
+import io.enmasse.systemtest.condition.SupportedInstallTypeCondition;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
