+++ systemtests/src/test/java/io/enmasse/systemtest/isolated/web/ChromeConsoleTest.java
@@ -7,11 +7,11 @@ package io.enmasse.systemtest.isolated.web;
 import io.enmasse.address.model.AddressSpaceBuilder;
 import io.enmasse.systemtest.bases.isolated.ITestIsolatedStandard;
 import io.enmasse.systemtest.bases.web.ConsoleTest;
-import io.enmasse.systemtest.messagingclients.ExternalClients;
+import io.enmasse.systemtest.annotations.ExternalClients;
 import io.enmasse.systemtest.model.addressplan.DestinationPlan;
 import io.enmasse.systemtest.model.addressspace.AddressSpacePlans;
 import io.enmasse.systemtest.model.addressspace.AddressSpaceType;
-import io.enmasse.systemtest.selenium.SeleniumChrome;
+import io.enmasse.systemtest.annotations.SeleniumChrome;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
