+++ systemtests/src/test/java/io/enmasse/systemtest/shared/brokered/web/ChromeWebSocketBrowserTest.java
@@ -8,7 +8,7 @@ import io.enmasse.address.model.AddressBuilder;
 import io.enmasse.systemtest.bases.shared.ITestSharedBrokered;
 import io.enmasse.systemtest.bases.web.WebSocketBrowserTest;
 import io.enmasse.systemtest.model.address.AddressType;
-import io.enmasse.systemtest.selenium.SeleniumChrome;
+import io.enmasse.systemtest.annotations.SeleniumChrome;
 import io.enmasse.systemtest.utils.AddressUtils;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
