+++ systemtests/src/test/java/io/enmasse/systemtest/shared/standard/web/FirefoxConsoleTest.java
@@ -9,10 +9,10 @@ import io.enmasse.systemtest.UserCredentials;
 import io.enmasse.systemtest.bases.shared.ITestSharedStandard;
 import io.enmasse.systemtest.bases.web.ConsoleTest;
 import io.enmasse.systemtest.info.TestInfo;
-import io.enmasse.systemtest.messagingclients.ExternalClients;
+import io.enmasse.systemtest.annotations.ExternalClients;
 import io.enmasse.systemtest.model.address.AddressType;
 import io.enmasse.systemtest.model.addressplan.DestinationPlan;
-import io.enmasse.systemtest.selenium.SeleniumFirefox;
+import io.enmasse.systemtest.annotations.SeleniumFirefox;
 import io.enmasse.systemtest.utils.AddressUtils;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
