+++ systemtests/src/test/java/io/enmasse/systemtest/shared/standard/web/FirefoxWebSocketBrowserTest.java
@@ -8,7 +8,7 @@ import io.enmasse.address.model.AddressBuilder;
 import io.enmasse.systemtest.bases.shared.ITestSharedStandard;
 import io.enmasse.systemtest.bases.web.WebSocketBrowserTest;
 import io.enmasse.systemtest.model.address.AddressType;
-import io.enmasse.systemtest.selenium.SeleniumFirefox;
+import io.enmasse.systemtest.annotations.SeleniumFirefox;
 import io.enmasse.systemtest.utils.AddressUtils;
 import org.junit.jupiter.api.Test;
